Move the proposed indicator method constant to scorecard_constant.js
